exception in controller

+++ src/news.controller.ts
@@ -2,11 +2,14 @@ import { NewsItems } from './interfaces/newsItems.interface';
 import { NewsDto } from './dto/news.dto';
 import { Controller, Get, Post, Body, Put, Param, Delete, HttpStatus, Res, HttpException, BadRequestException } from '@nestjs/common'
 import { NewsService } from './news.service'
+import { NotFoundExceptionForService } from './notFonud.exception';
 
 
 @Controller('news')
 export class NewsController {
   constructor(private readonly newsService: NewsService) { }
+
+
 
   @Post()
   create(@Body() newsDto: NewsDto): Promise<NewsItems> {
@@ -20,7 +23,12 @@ export class NewsController {
 
   @Get(':id')
   async getNewsById(@Param('id') id: string): Promise<NewsItems | null> {
-    return this.newsService.findById(id);
+    const news: NewsItems = await this.newsService.findById(id)
+    if (news === undefined) {
+      throw new NotFoundExceptionForService(id)
+    } else {
+      return news;
+    }
   }
 
   @Put(':id')
@@ -28,12 +36,22 @@ export class NewsController {
     @Param('id') id: string,
     @Body() newsDto: NewsDto
   ): Promise<NewsItems | null> {
-    return this.newsService.update(id, newsDto);
+    const news: NewsItems = await this.newsService.update(id, newsDto)
+    if (news === undefined) {
+      throw new NotFoundExceptionForService(id)
+    } else {
+      return news;
+    }
   }
 
   @Delete(':id')
   async remove(@Param('id') id: string): Promise<void | null> {
-    return this.newsService.remove(id);
+    // const news: NewsItems = await this.newsService.remove(id);
+    if (this.newsService.remove(id) === undefined) {
+      throw new NotFoundExceptionForService(id)
+    } else {
+      return this.newsService.remove(id);
+    }
   }
 
 }
